feat: criacao de todas entidades base para a classificacao

+++ backend/src/main/java/com/example/fiscos/dto/nfeApi/SummaryNFeDTO.java
@@ -1,5 +1,7 @@
 package com.example.fiscos.dto.nfeApi;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import jakarta.validation.constraints.*;
@@ -18,11 +20,11 @@ public class SummaryNFeDTO {
 
     @PositiveOrZero(message = "Valor total deve ser zero ou positivo.")
     @JsonProperty("totalAmount")
-    private double totalAmount;
+    private BigDecimal totalAmount;
 
     @PositiveOrZero(message = "Valor pago deve ser zero ou positivo.")
     @JsonProperty("amountPaid")
-    private double amountPaid;
+    private BigDecimal amountPaid;
 
     @NotBlank(message = "Método de pagamento não pode estar vazio.")
     @JsonProperty("paymentMethod")
