+++ backend/src/main/java/com/example/fiscos/dto/nfeApi/TaxNFeDTO.java
@@ -1,5 +1,7 @@
 package com.example.fiscos.dto.nfeApi;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import jakarta.validation.constraints.*;
@@ -14,13 +16,13 @@ public class TaxNFeDTO {
 
     @PositiveOrZero(message = "Valor total do imposto deve ser zero ou positivo.")
     @JsonProperty("totalTaxAmount")
-    private double totalTaxAmount;
+    private BigDecimal totalTaxAmount;
 
     @PositiveOrZero(message = "Valor do imposto federal deve ser zero ou positivo.")
     @JsonProperty("federal")
-    private double federal;
+    private BigDecimal federal;
 
     @PositiveOrZero(message = "Valor do imposto estadual deve ser zero ou positivo.")
     @JsonProperty("state")
-    private double state;
+    private BigDecimal state;
 }
